change double to int

+++ lib/data/models/car.dart
@@ -1,9 +1,9 @@
 class Car {
   final String? image;
   final String model;
-  final double distance;
-  final double fuelCapacity;
-  final double pricePerHour;
+  final int distance;
+  final int fuelCapacity;
+  final int pricePerHour;
 
   Car( {this.image,required this.model, required this.distance, required this.fuelCapacity, required this.pricePerHour});
